OCD-38: Incorporate Table section into Query
