+++ client/src/js/tracks.js
@@ -3,9 +3,7 @@ import React, { Component } from 'react';
 class Tracks extends Component { 
   render() {
     return (
-      <div>
-        <iframe src="http://172.20.4.59:3002" title="JBrowse" width="90%" height="400em" />
-      </div>
+      <iframe src="http://172.20.4.59:3002" title="JBrowse" width="90%" height="100%" />
     );
   }
 }
